refactor: move quorum rotation related code to QuorumRotationState
