+++ core/src/main/java/org/bitcoinj/evolution/SimplifiedMasternodeListManager.java
@@ -13,14 +13,10 @@ import org.bitcoinj.evolution.listeners.MasternodeListDownloadedListener;
 import org.bitcoinj.quorums.GetQuorumRotationInfo;
 import org.bitcoinj.quorums.LLMQParameters;
 import org.bitcoinj.quorums.LLMQUtils;
-import org.bitcoinj.quorums.PreviousQuorumQuarters;
-import org.bitcoinj.quorums.Quorum;
 import org.bitcoinj.quorums.QuorumManager;
 import org.bitcoinj.quorums.QuorumRotationInfo;
-import org.bitcoinj.quorums.QuorumSnapshot;
 import org.bitcoinj.quorums.QuorumSnapshotManager;
 import org.bitcoinj.quorums.SigningManager;
-import org.bitcoinj.quorums.SnapshotSkipMode;
 import org.bitcoinj.store.BlockStoreException;
 import org.bitcoinj.utils.Pair;
 import org.bitcoinj.utils.Threading;
@@ -33,9 +29,7 @@ import javax.annotation.concurrent.GuardedBy;
 import java.io.*;
 import java.nio.ByteBuffer;
 import java.util.*;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.locks.ReentrantLock;
-import java.util.stream.Collectors;
 
 public class SimplifiedMasternodeListManager extends AbstractManager {
     private static final Logger log = LoggerFactory.getLogger(SimplifiedMasternodeListManager.class);
@@ -127,32 +121,13 @@ public class SimplifiedMasternodeListManager extends AbstractManager {
     QuorumSnapshotManager quorumSnapshotManager;
     QuorumManager quorumManager;
 
-    QuorumSnapshot quorumSnapshotAtHMinusC;
-    QuorumSnapshot quorumSnapshotAtHMinus2C;
-    QuorumSnapshot quorumSnapshotAtHMinus3C;
-    SimplifiedMasternodeList mnListTip;
-    SimplifiedMasternodeList mnListAtHMinusC;
-    SimplifiedMasternodeList mnListAtHMinus2C;
-    SimplifiedMasternodeList mnListAtHMinus3C;
-    SimplifiedQuorumList quorumListTip;
-    SimplifiedQuorumList quorumListAtHMinusC;
-    SimplifiedQuorumList quorumListAtHMinus2C;
-    SimplifiedQuorumList quorumListAtHMinus3C;
+    QuorumRotationState quorumRotationState;
 
     public SimplifiedMasternodeListManager(Context context) {
         super(context);
         tipBlockHash = params.getGenesisBlock().getHash();
         mnList = new SimplifiedMasternodeList(context.getParams());
         quorumList = new SimplifiedQuorumList(context.getParams());
-
-        mnListTip = new SimplifiedMasternodeList(context.getParams());
-        mnListAtHMinusC = new SimplifiedMasternodeList(context.getParams());
-        mnListAtHMinus2C = new SimplifiedMasternodeList(context.getParams());
-        mnListAtHMinus3C = new SimplifiedMasternodeList(context.getParams());
-        quorumListTip = new SimplifiedQuorumList(context.getParams());
-        quorumListAtHMinusC = new SimplifiedQuorumList(context.getParams());
-        quorumListAtHMinus2C = new SimplifiedQuorumList(context.getParams());
-        quorumListAtHMinus3C = new SimplifiedQuorumList(context.getParams());
 
         lastRequestTime = 0;
         waitingForMNListDiff = false;
@@ -165,6 +140,8 @@ public class SimplifiedMasternodeListManager extends AbstractManager {
         requiresLoadingFromFile = true;
         lastRequestMessage = new GetSimplifiedMasternodeListDiff(Sha256Hash.ZERO_HASH, Sha256Hash.ZERO_HASH);
         initChainTipSyncComplete = !context.getSyncFlags().contains(MasternodeSync.SYNC_FLAGS.SYNC_HEADERS_MN_LIST_FIRST);
+
+        quorumRotationState = new QuorumRotationState(context);
     }
 
     @Override
@@ -396,6 +373,7 @@ public class SimplifiedMasternodeListManager extends AbstractManager {
     }
 
     public void processQuorumRotationInfo(@Nullable Peer peer, QuorumRotationInfo quorumRotationInfo, boolean isLoadingBootStrap) {
+
         StoredBlock block;
         StoredBlock blockTip;
         StoredBlock blockMinusC;
@@ -411,89 +389,9 @@ public class SimplifiedMasternodeListManager extends AbstractManager {
         log.info("processing quorumrotationinfo between : {} & {}; {}", mnList.getHeight(), newHeight, quorumRotationInfo);
         lock.lock();
         try {
-            block = chain.getBlockStore().get(quorumRotationInfo.getMnListDiffTip().blockHash);
-            blockTip = chain.getBlockStore().get((int)quorumRotationInfo.getCreationHeight());
-            blockMinusC = chain.getBlockStore().get(quorumRotationInfo.getMnListDiffAtHMinusC().blockHash);
-            blockMinus2C = chain.getBlockStore().get(quorumRotationInfo.getMnListDiffAtHMinus2C().blockHash);
-            blockMinus3C = chain.getBlockStore().get(quorumRotationInfo.getMnListDiffAtHMinus3C().blockHash);
+            quorumRotationState.applyDiff(peer, headersChain, blockChain, quorumRotationInfo, isLoadingBootStrap);
 
-            if(!isLoadingBootStrap && block.getHeight() != newHeight)
-                throw new ProtocolException("mnlistdiff blockhash (height="+block.getHeight()+" doesn't match coinbase blockheight: " + newHeight);
 
-            watchMNList.start();
-            if (peer != null && isSyncingHeadersFirst) peer.queueMasternodeListDownloadedListeners(MasternodeListDownloadedListener.Stage.Processing, quorumRotationInfo.getMnListDiffTip());
-
-            SimplifiedMasternodeList newMNListTip = mnListTip.applyDiff(quorumRotationInfo.getMnListDiffTip());
-            SimplifiedMasternodeList newNMListAtHMinusC = mnListAtHMinusC.applyDiff(quorumRotationInfo.getMnListDiffAtHMinusC());
-            SimplifiedMasternodeList newNMListAtHMinus2C = mnListAtHMinus2C.applyDiff(quorumRotationInfo.getMnListDiffAtHMinus2C());
-            SimplifiedMasternodeList newNMListAtHMinus3C = mnListAtHMinus3C.applyDiff(quorumRotationInfo.getMnListDiffAtHMinus3C());
-            if(context.masternodeSync.hasVerifyFlag(MasternodeSync.VERIFY_FLAGS.MNLISTDIFF_MNLIST)) {
-                newMNListTip.verify(quorumRotationInfo.getMnListDiffTip().coinBaseTx, quorumRotationInfo.getMnListDiffTip(), mnListTip);
-                newNMListAtHMinusC.verify(quorumRotationInfo.getMnListDiffAtHMinusC().coinBaseTx, quorumRotationInfo.getMnListDiffAtHMinusC(), mnListAtHMinusC);
-                newNMListAtHMinus2C.verify(quorumRotationInfo.getMnListDiffAtHMinus2C().coinBaseTx, quorumRotationInfo.getMnListDiffAtHMinus2C(), mnListAtHMinus2C);
-                newNMListAtHMinus3C.verify(quorumRotationInfo.getMnListDiffAtHMinus3C().coinBaseTx, quorumRotationInfo.getMnListDiffAtHMinus3C(), mnListAtHMinus3C);
-            }
-
-            if (peer != null && isSyncingHeadersFirst)
-                peer.queueMasternodeListDownloadedListeners(MasternodeListDownloadedListener.Stage.ProcessedMasternodes, quorumRotationInfo.getMnListDiffTip());
-
-            newMNListTip.setBlock(blockTip, block != null && block.getHeader().getPrevBlockHash().equals(quorumRotationInfo.getMnListDiffTip().prevBlockHash));
-            newNMListAtHMinusC.setBlock(blockMinusC, block != null && blockMinusC.getHeader().getPrevBlockHash().equals(quorumRotationInfo.getMnListDiffAtHMinusC().prevBlockHash));
-            newNMListAtHMinus2C.setBlock(blockMinus2C, block != null && blockMinus2C.getHeader().getPrevBlockHash().equals(quorumRotationInfo.getMnListDiffAtHMinus2C().prevBlockHash));
-            newNMListAtHMinus3C.setBlock(blockMinus3C, block != null && blockMinus3C.getHeader().getPrevBlockHash().equals(quorumRotationInfo.getMnListDiffAtHMinus3C().prevBlockHash));
-
-            mnListsCache.clear();
-            mnListsCache.put(blockTip.getHeader().getHash(), newMNListTip);
-            mnListsCache.put(newNMListAtHMinusC.getBlockHash(), newNMListAtHMinusC);
-            mnListsCache.put(newNMListAtHMinus2C.getBlockHash(), newNMListAtHMinus2C);
-            mnListsCache.put(newNMListAtHMinus3C.getBlockHash(), newNMListAtHMinus3C);
-
-            quorumSnapshotAtHMinusC = quorumRotationInfo.getQuorumSnapshotAtHMinusC();
-            quorumSnapshotAtHMinus2C = quorumRotationInfo.getQuorumSnapshotAtHMinus2C();
-            quorumSnapshotAtHMinus3C = quorumRotationInfo.getQuorumSnapshotAtHMinus3C();
-
-            // now calculate quorums?
-
-            //SimplifiedQuorumList newQuorumListTip = quorumListTip;
-            //SimplifiedQuorumList newQuorumListTipAtHMinusC = quorumListAtHMinusC;
-            //SimplifiedQuorumList newQuorumListTipAtHMinus2C = quorumListAtHMinus2C;
-            //SimplifiedQuorumList newQuorumListTipAtHMinus3C = quorumListAtHMinus3C;
-
-            watchQuorums.start();
-
-            SimplifiedQuorumList newQuorumListTip = quorumListTip.applyDiff(quorumRotationInfo.getMnListDiffTip(), isLoadingBootStrap, chain);
-            SimplifiedQuorumList newQuorumListTipAtHMinusC = quorumListAtHMinusC.applyDiff(quorumRotationInfo.getMnListDiffAtHMinusC(), isLoadingBootStrap, chain);
-            SimplifiedQuorumList newQuorumListTipAtHMinus2C = quorumListAtHMinus2C.applyDiff(quorumRotationInfo.getMnListDiffAtHMinus2C(), isLoadingBootStrap, chain);
-            SimplifiedQuorumList newQuorumListTipAtHMinus3C = quorumListAtHMinus3C.applyDiff(quorumRotationInfo.getMnListDiffAtHMinus3C(), isLoadingBootStrap, chain);
-            if(context.masternodeSync.hasVerifyFlag(MasternodeSync.VERIFY_FLAGS.MNLISTDIFF_QUORUM)) {
-                newQuorumListTip.verify(quorumRotationInfo.getMnListDiffTip().coinBaseTx, quorumRotationInfo.getMnListDiffTip(), quorumListTip, newMNListTip);
-                newQuorumListTipAtHMinusC.verify(quorumRotationInfo.getMnListDiffAtHMinusC().coinBaseTx, quorumRotationInfo.getMnListDiffAtHMinusC(), quorumListAtHMinusC, newNMListAtHMinusC);
-                newQuorumListTipAtHMinus2C.verify(quorumRotationInfo.getMnListDiffAtHMinus2C().coinBaseTx, quorumRotationInfo.getMnListDiffAtHMinus2C(), quorumListAtHMinus2C, newNMListAtHMinus2C);
-                newQuorumListTipAtHMinus3C.verify(quorumRotationInfo.getMnListDiffAtHMinus3C().coinBaseTx, quorumRotationInfo.getMnListDiffAtHMinus3C(), quorumListAtHMinus3C, newNMListAtHMinus3C);
-            }
-
-            if (peer != null && isSyncingHeadersFirst)
-                peer.queueMasternodeListDownloadedListeners(MasternodeListDownloadedListener.Stage.ProcessedQuorums, quorumRotationInfo.getMnListDiffTip());
-
-            quorumsCache.clear();
-            quorumsCache.put(blockTip.getHeader().getHash(), newQuorumListTip);
-            quorumsCache.put(newQuorumListTipAtHMinusC.getBlockHash(), newQuorumListTipAtHMinusC);
-            quorumsCache.put(newQuorumListTipAtHMinus2C.getBlockHash(), newQuorumListTipAtHMinus2C);
-            quorumsCache.put(newQuorumListTipAtHMinus3C.getBlockHash(), newQuorumListTipAtHMinus3C);
-
-            mnList = newMNListTip;
-            mnListTip = newMNListTip;
-            mnListAtHMinusC = newNMListAtHMinusC;
-            mnListAtHMinus2C = newNMListAtHMinus2C;
-            mnListAtHMinus3C = newNMListAtHMinus3C;
-
-            quorumList = newQuorumListTip;
-            quorumListTip = newQuorumListTip;
-            quorumListAtHMinusC = newQuorumListTipAtHMinusC;
-            quorumListAtHMinus2C = newQuorumListTipAtHMinus2C;
-            quorumListAtHMinus3C = newQuorumListTipAtHMinus3C;
-
-            //quorumList = newQuorumList;
             log.info(this.toString());
             unCache();
             failedAttempts = 0;
@@ -510,9 +408,9 @@ public class SimplifiedMasternodeListManager extends AbstractManager {
             if(quorumRotationInfo.getMnListDiffTip().hasChanges() || pendingBlocks.size() < MAX_CACHE_SIZE || saveOptions == SaveOptions.SAVE_EVERY_BLOCK)
                 save();
 
-            log.info("getAllQuorums: {}", getAllQuorumMembers(params.getLlmqForInstantSend(), blockTip.getHeader().getHash()).size());
-            ArrayList<Quorum> list = context.quorumManager.scanQuorums(params.getLlmqForInstantSend(), LLMQParameters.fromType(params.getLlmqForInstantSend()).getSigningActiveQuorumCount());
-            log.info("quorum list: {}", list.toString());
+            // log.info("getAllQuorums: {}", getAllQuorumMembers(params.getLlmqForInstantSend(), blockTip.getHeader().getHash()).size());
+            // ArrayList<Quorum> list = context.quorumManager.scanQuorums(params.getLlmqForInstantSend(), LLMQParameters.fromType(params.getLlmqForInstantSend()).getSigningActiveQuorumCount());
+            // log.info("quorum list: {}", list.toString());
         } catch(MasternodeListDiffException x) {
             //we already have this mnlistdiff or doesn't match our current tipBlockHash
             if(mnList.getBlockHash().equals(quorumRotationInfo.getMnListDiffTip().blockHash)) {
@@ -724,8 +622,6 @@ public class SimplifiedMasternodeListManager extends AbstractManager {
                     requestNextMNListDiff();
                     return;
                 }
-                //if (lastRequestTime + WAIT_GETMNLISTDIFF < Utils.currentTimeMillis())
-                //    return;
                 return;
             } else if (lastRequestTime + WAIT_GETMNLISTDIFF > Utils.currentTimeMillis() ||
                     blockChain.getChainHead().getHeader().getTimeSeconds() < Utils.currentTimeSeconds() - timePeriod) {
@@ -783,7 +679,7 @@ public class SimplifiedMasternodeListManager extends AbstractManager {
                     // get from the last one
                 //    msg = new GetQuorumRotationInfo(params, 4, Lists.newArrayList(mnListTip.getBlockHash(), mnListAtHMinusC.getBlockHash(), mnListAtHMinus2C.getBlockHash(), mnListAtHMinus3C.getBlockHash()),endBlock.getHeader().getHash());
                 //}
-                lastRequestHash = mnListTip.getBlockHash();
+                lastRequestHash = quorumRotationState.getMnListTip().getBlockHash();
                 downloadPeer.sendMessage(msg);
             }
             lastRequestTime = Utils.currentTimeMillis();
@@ -860,13 +756,7 @@ public class SimplifiedMasternodeListManager extends AbstractManager {
                     log.info("requesting mnlistdiff from {} to {}; \n  From {}\n To {}", mnList.getHeight(), nextBlock.getHeight(), mnList.getBlockHash(), nextBlock.getHeader().getHash());
                     GetSimplifiedMasternodeListDiff requestMessage = new GetSimplifiedMasternodeListDiff(mnList.getBlockHash(), nextBlock.getHeader().getHash());
 
-                    GetQuorumRotationInfo requestMessage2 = new GetQuorumRotationInfo(params, 4,
-                            Lists.newArrayList(
-                                    mnListTip.getBlockHash(),
-                                    mnListAtHMinusC.getBlockHash(),
-                                    mnListAtHMinus2C.getBlockHash(),
-                                    mnListAtHMinus3C.getBlockHash()),
-                            nextBlock.getHeader().getHash());
+                    GetQuorumRotationInfo requestMessage2 = quorumRotationState.getQuorumRotationInfoRequest(nextBlock);
 
                     if(requestMessage.equals(lastRequestMessage)) {
                         log.info("request for mnlistdiff is the same as the last request");
@@ -1071,67 +961,10 @@ public class SimplifiedMasternodeListManager extends AbstractManager {
 
     public ArrayList<Masternode> getAllQuorumMembers(LLMQParameters.LLMQType llmqType, Sha256Hash blockHash)
     {
-        lock.lock();
+
         try {
             if (LLMQUtils.isQuorumRotationEnabled(context, params, llmqType)) {
-                ArrayList<Masternode> quorumMembers;
-                if (mapQuorumMembers.isEmpty()) {
-                    initQuorumsCache(mapQuorumMembers);
-                }
-                HashMap<Sha256Hash, ArrayList<Masternode>> mapByType = mapQuorumMembers.get(llmqType);
-                if (mapByType != null) {
-                    quorumMembers = mapByType.get(blockHash);
-                    if (quorumMembers != null)
-                        return quorumMembers;
-                }
-
-                StoredBlock quorumBaseBlock = blockChain.getBlockStore().get(blockHash);
-                if (mapIndexedQuorumMembers.isEmpty()) {
-                    initIndexedQuorumsCache(mapIndexedQuorumMembers);
-                }
-
-                /*
-                 * Quorums created with rotation are now created in a different way. All signingActiveQuorumCount are created during the period of dkgInterval.
-                 * But they are not created exactly in the same block, they are spread overtime: one quorum in each block until all signingActiveQuorumCount are created.
-                 * The new concept of quorumIndex is introduced in order to identify them.
-                 * In every dkgInterval blocks (also called CycleQuorumBaseBlock), the spread quorum creation starts like this:
-                 * For quorumIndex = 0 : signingActiveQuorumCount
-                 * Quorum Q with quorumIndex is created at height CycleQuorumBaseBlock + quorumIndex
-                 */
-
-                int quorumIndex = quorumBaseBlock.getHeight() % LLMQParameters.fromType(llmqType).getDkgInterval();
-                int cycleQuorumBaseHeight = quorumBaseBlock.getHeight() - quorumIndex;
-                final StoredBlock cycleQuorumBaseBlock = blockChain.getBlockStore().get(cycleQuorumBaseHeight);
-
-                /*
-                 * Since mapQuorumMembers stores Quorum members per block hash, and we don't know yet the block hashes of blocks for all quorumIndexes (since these blocks are not created yet)
-                 * We store them in a second cache mapIndexedQuorumMembers which stores them by {CycleQuorumBaseBlockHash, quorumIndex}
-                 */
-
-                HashMap<Pair<Sha256Hash, Integer>, ArrayList<Masternode>> mapByTypeIndexed = mapIndexedQuorumMembers.get(llmqType);
-                if (mapByType != null) {
-                    quorumMembers = mapByTypeIndexed.get(new Pair<>(cycleQuorumBaseBlock.getHeader().getHash(), quorumIndex));
-                    if (quorumMembers != null) {
-                        mapQuorumMembers.get(llmqType).put(cycleQuorumBaseBlock.getHeader().getHash(), quorumMembers);
-                        /*
-                         * We also need to store which quorum block hash corresponds to which quorumIndex
-                         */
-                        //quorumManager.setQuorumIndexQuorumHash(llmqType, quorumBaseBlock.getHeader().getHash(), quorumIndex);
-                        return quorumMembers;
-                    }
-                }
-
-                ArrayList<ArrayList<Masternode>> q = computeQuorumMembersByQuarterRotation(llmqType, cycleQuorumBaseBlock);
-                for (int i = 0; i < q.size(); ++i) {
-                    mapIndexedQuorumMembers.get(llmqType).put(new Pair<>(cycleQuorumBaseBlock.getHeader().getHash(), i), q.get(i));
-                }
-
-                quorumMembers = q.get(0);
-
-                mapQuorumMembers.get(llmqType).put(quorumBaseBlock.getHeader().getHash(), quorumMembers);
-                //quorumManager.setQuorumIndexQuorumHash(llmqType, quorumBaseBlock.getHeader().getHash(), quorumIndex);
-
-                return quorumMembers;
+                return quorumRotationState.getAllQuorumMembers(llmqType, blockHash);
             } else {
                 LLMQParameters llmqParameters = params.getLlmqs().get(llmqType);
                 SimplifiedMasternodeList allMns = getListForBlock(blockHash);
@@ -1141,346 +974,9 @@ public class SimplifiedMasternodeListManager extends AbstractManager {
                 }
                 return null;
             }
-        } catch (BlockStoreException x) {
-            // we are in deep trouble
-            throw new RuntimeException(x);
         } finally {
             lock.unlock();
         }
-    }
-
-    private ArrayList<ArrayList<Masternode>> computeQuorumMembersByQuarterRotation(LLMQParameters.LLMQType llmqType, StoredBlock quorumBaseBlock) throws BlockStoreException {
-        final LLMQParameters llmqParameters = LLMQParameters.fromType(llmqType);
-
-        final int cycleLength = llmqParameters.getDkgInterval();
-
-        final StoredBlock blockHMinusC = quorumBaseBlock.getAncestor(blockChain.getBlockStore(), quorumBaseBlock.getHeight() - cycleLength);
-        final StoredBlock pBlockHMinus2C = quorumBaseBlock.getAncestor(blockChain.getBlockStore(),quorumBaseBlock.getHeight() - 2 * cycleLength);
-        final StoredBlock pBlockHMinus3C = quorumBaseBlock.getAncestor(blockChain.getBlockStore(),quorumBaseBlock.getHeight() - 3 * cycleLength);
-
-        PreviousQuorumQuarters previousQuarters = getPreviousQuorumQuarterMembers(llmqParameters, blockHMinusC, pBlockHMinus2C, pBlockHMinus3C);
-
-        ArrayList<ArrayList<Masternode>> quorumMembers = Lists.newArrayListWithCapacity(llmqParameters.getSigningActiveQuorumCount());
-        for (int i = 0; i < llmqParameters.getSigningActiveQuorumCount(); ++i) {
-            quorumMembers.add(Lists.newArrayList());
-        }
-
-        ArrayList<ArrayList<SimplifiedMasternodeListEntry>> newQuarterMembers = buildNewQuorumQuarterMembers(llmqParameters, quorumBaseBlock, previousQuarters);
-
-        for (int i = 0; i < llmqParameters.getSigningActiveQuorumCount(); ++i) {
-            for (SimplifiedMasternodeListEntry m: previousQuarters.quarterHMinus3C.get(i)) {
-                for (int j = 0; j < quorumMembers.get(i).size(); ++j) {
-                    if (m.equals(quorumMembers.get(i).get(j))) {
-                        log.info("{} is already in the list", m);
-                    }
-                }
-                quorumMembers.get(i).add(m);
-            }
-            for (SimplifiedMasternodeListEntry m: previousQuarters.quarterHMinus2C.get(i)) {
-                for (int j = 0; j < quorumMembers.get(i).size(); ++j) {
-                    if (m.equals(quorumMembers.get(i).get(j))) {
-                        log.info("{} is already in the list", m);
-                    }
-                }
-                quorumMembers.get(i).add(m);
-            }
-            for (SimplifiedMasternodeListEntry m: previousQuarters.quarterHMinusC.get(i)) {
-                for (int j = 0; j < quorumMembers.get(i).size(); ++j) {
-                    if (m.equals(quorumMembers.get(i).get(j))) {
-                        log.info("{} is already in the list", m);
-                    }
-                }
-                quorumMembers.get(i).add(m);
-            }
-            for (SimplifiedMasternodeListEntry m: newQuarterMembers.get(i)) {
-                for (int j = 0; j < quorumMembers.get(i).size(); ++j) {
-                    if (m.equals(quorumMembers.get(i).get(j))) {
-                        log.info("{} is already in the list", m);
-                    }
-                }
-                quorumMembers.get(i).add(m);
-            }
-        }
-
-        return quorumMembers;
-    }
-
-    private ArrayList<ArrayList<SimplifiedMasternodeListEntry>> buildNewQuorumQuarterMembers(LLMQParameters llmqParameters, StoredBlock quorumBaseBlock, PreviousQuorumQuarters previousQuarters)
-    {
-
-        int nQuorums = llmqParameters.getSigningActiveQuorumCount();
-        ArrayList<ArrayList<SimplifiedMasternodeListEntry>> quarterQuorumMembers = new ArrayList<>(nQuorums);
-        for (int i = 0; i < llmqParameters.getSigningActiveQuorumCount(); ++i) {
-            quarterQuorumMembers.add(Lists.newArrayList());
-        }
-        int quorumSize = llmqParameters.getSize();
-        int quarterSize = quorumSize / 4;
-
-        Sha256Hash modifier = LLMQUtils.buildLLMQBlockHash(llmqParameters.getType(), quorumBaseBlock.getHeader().getHash());
-        SimplifiedMasternodeList allMns = getListForBlock(quorumBaseBlock.getHeader().getHash());
-        if (allMns.getAllMNsCount() < quarterSize)
-            return quarterQuorumMembers;
-
-        SimplifiedMasternodeList MnsUsedAtH = new SimplifiedMasternodeList(params);
-        SimplifiedMasternodeList MnsNotUsedAtH = new SimplifiedMasternodeList(params);
-        ArrayList<SimplifiedMasternodeList> MnsUsedAtHIndex = Lists.newArrayListWithCapacity(llmqParameters.getSigningActiveQuorumCount());
-        for (int i = 0; i < llmqParameters.getSigningActiveQuorumCount(); ++i) {
-            MnsUsedAtHIndex.add(new SimplifiedMasternodeList(params));
-        }
-
-        for (int i = 0; i < llmqParameters.getSigningActiveQuorumCount(); ++i) {
-            for (SimplifiedMasternodeListEntry mn: previousQuarters.quarterHMinusC.get(i)) {
-                MnsUsedAtH.addMN(mn);
-                MnsUsedAtHIndex.get(i).addMN(mn);
-            }
-            for (SimplifiedMasternodeListEntry mn: previousQuarters.quarterHMinus2C.get(i)) {
-                MnsUsedAtH.addMN(mn);
-                MnsUsedAtHIndex.get(i).addMN(mn);
-            }
-            for (SimplifiedMasternodeListEntry mn: previousQuarters.quarterHMinus3C.get(i)) {
-                MnsUsedAtH.addMN(mn);
-                MnsUsedAtHIndex.get(i).addMN(mn);
-            }
-        }
-
-        allMns.forEachMN(true, mn -> {
-            if (!MnsUsedAtH.containsMN(mn.getProTxHash())){
-                MnsNotUsedAtH.addMN(mn);
-            }
-        });
-
-        ArrayList<Masternode> sortedMnsUsedAtH = MnsUsedAtH.calculateQuorum(MnsUsedAtH.getAllMNsCount(), modifier);
-        ArrayList<Masternode> sortedMnsNotUsedAtH = MnsNotUsedAtH.calculateQuorum(MnsNotUsedAtH.getAllMNsCount(), modifier);
-        ArrayList<Masternode> sortedCombinedMnsList = new ArrayList<>(sortedMnsNotUsedAtH);
-        sortedCombinedMnsList.addAll(sortedMnsUsedAtH);
-
-        ArrayList<Integer> skipList = Lists.newArrayList();
-        int firstSkippedIndex = 0;
-        int idx = 0;
-        for (int i = 0; i < llmqParameters.getSigningActiveQuorumCount(); ++i) {
-            while (quarterQuorumMembers.get(i).size() < quarterSize) {
-                Masternode mn = sortedCombinedMnsList.get(idx);
-                if (!MnsUsedAtHIndex.get(i).containsMN(mn.getProTxHash())) {
-                    quarterQuorumMembers.get(i).add((SimplifiedMasternodeListEntry) mn);
-                } else {
-                    if (firstSkippedIndex == 0) {
-                        firstSkippedIndex = idx;
-                        skipList.add(idx);
-                    } else {
-                        skipList.add(idx - firstSkippedIndex);
-                    }
-                }
-                idx++;
-                if (idx == sortedCombinedMnsList.size()) {
-                    idx = 0;
-                }
-            }
-        }
-
-        // maybe we don't need to do this as a client, lets check it
-        QuorumSnapshot quorumSnapshot = buildQuorumSnapshot(llmqParameters, allMns, MnsUsedAtH, sortedCombinedMnsList, skipList);
-        log.info("quorumSnapshot = {}", quorumSnapshot);
-
-        // TODO: Do we need this?
-        // quorumSnapshotManager.storeSnapshotForBlock(llmqParameters.getType(), quorumBaseBlock, quorumSnapshot);
-
-        return quarterQuorumMembers;
-    }
-
-
-
-    private QuorumSnapshot buildQuorumSnapshot(LLMQParameters llmqParameters, SimplifiedMasternodeList mnAtH, SimplifiedMasternodeList mnUsedAtH, ArrayList<Masternode> sortedCombinedMns, ArrayList<Integer> skipList)
-    {
-        QuorumSnapshot quorumSnapshot = new QuorumSnapshot(mnAtH.getAllMNsCount());
-
-        AtomicInteger index = new AtomicInteger();
-        mnAtH.forEachMN(true, mn -> {
-            if (mnUsedAtH.containsMN(mn.getProTxHash())){
-                quorumSnapshot.setActiveQuorumMember(index.get(), true);
-            }
-            index.getAndIncrement();
-        });
-
-        // TODO: do we need this?
-        // buildQuorumSnapshotSkipList(llmqParameters, mnUsedAtH, sortedCombinedMns, quorumSnapshot);
-        if (skipList.isEmpty()) {
-            quorumSnapshot.setSkipListMode(SnapshotSkipMode.MODE_NO_SKIPPING);
-            quorumSnapshot.getSkipList().clear();
-        } else {
-            quorumSnapshot.setSkipListMode(SnapshotSkipMode.MODE_SKIPPING_ENTRIES);
-            quorumSnapshot.setSkipList(skipList);
-        }
-        return quorumSnapshot;
-    }
-
-
-    void buildQuorumSnapshotSkipList(LLMQParameters llmqParameters, SimplifiedMasternodeList mnUsedAtH, ArrayList<Masternode> sortedCombinedMns, QuorumSnapshot quorumSnapshot)
-    {
-        if (mnUsedAtH.getAllMNsCount() == 0) {
-            quorumSnapshot.setSkipListMode(SnapshotSkipMode.MODE_NO_SKIPPING);
-            quorumSnapshot.getSkipList().clear();
-        } else if (mnUsedAtH.getAllMNsCount() < sortedCombinedMns.size() / 2) {
-            quorumSnapshot.setSkipListMode(SnapshotSkipMode.MODE_SKIPPING_ENTRIES);
-
-            int first_entry_index = 0;
-            int index = 0;
-
-            for (Masternode mn : sortedCombinedMns) {
-                if (mnUsedAtH.containsMN(sortedCombinedMns.get(index).getProTxHash())) {
-                    if (first_entry_index == 0) {
-                        first_entry_index = index;
-                        quorumSnapshot.getSkipList().add(index);
-                    } else
-                        quorumSnapshot.getSkipList().add(index - first_entry_index);
-                }
-                index++;
-            }
-        }
-        else {
-            //Mode 2: Non-Skipping entries
-            quorumSnapshot.setSkipListMode(SnapshotSkipMode.MODE_NO_SKIPPING_ENTRIES);
-
-            int first_entry_index = 0;
-            int index = 0;
-
-            for (Masternode mn : sortedCombinedMns) {
-                if (!mnUsedAtH.containsMN(sortedCombinedMns.get(index).getProTxHash())) {
-                    if (first_entry_index == 0) {
-                        first_entry_index = index;
-                        quorumSnapshot.getSkipList().add(index);
-                    } else
-                        quorumSnapshot.getSkipList().add(index - first_entry_index);
-                }
-                index++;
-            }
-        }
-    }
-
-    PreviousQuorumQuarters getPreviousQuorumQuarterMembers(LLMQParameters llmqParameters, StoredBlock blockHMinusC, StoredBlock blockHMinus2C, StoredBlock blockHMinus3C) {
-        PreviousQuorumQuarters quarters = new PreviousQuorumQuarters();
-
-        QuorumSnapshot quSnapshotHMinusC = quorumSnapshotAtHMinusC;
-        if (quSnapshotHMinusC != null){
-
-            quarters.quarterHMinusC = getQuorumQuarterMembersBySnapshot(llmqParameters, blockHMinusC, quSnapshotHMinusC);
-
-            QuorumSnapshot quSnapshotHMinus2C = quorumSnapshotAtHMinus2C;
-            if (quSnapshotHMinus2C != null) {
-                quarters.quarterHMinus2C = getQuorumQuarterMembersBySnapshot(llmqParameters, blockHMinus2C, quSnapshotHMinus2C);
-
-                QuorumSnapshot quSnapshotHMinus3C = quorumSnapshotAtHMinus3C;
-                if (quSnapshotHMinus3C != null) {
-                    quarters.quarterHMinus3C = getQuorumQuarterMembersBySnapshot(llmqParameters, blockHMinus3C, quSnapshotHMinus3C);
-                }
-            }
-        }
-
-        return quarters;
-    }
-
-    ArrayList<ArrayList<SimplifiedMasternodeListEntry>> getQuorumQuarterMembersBySnapshot(LLMQParameters llmqParameters, StoredBlock quorumBaseBlock, QuorumSnapshot snapshot)
-    {
-        int numQuorums = llmqParameters.getSigningActiveQuorumCount();
-        int quorumSize = llmqParameters.getSize();
-        int quarterSize = quorumSize / 4;
-
-        ArrayList<ArrayList<SimplifiedMasternodeListEntry>> quarterQuorumMembers = Lists.newArrayListWithCapacity(numQuorums);
-        for (int i = 0; i < numQuorums; ++i) {
-            quarterQuorumMembers.add(Lists.newArrayList());
-        }
-
-        Sha256Hash modifier = LLMQUtils.buildLLMQBlockHash(llmqParameters.getType(), quorumBaseBlock.getHeader().getHash());
-
-        Pair<SimplifiedMasternodeList, SimplifiedMasternodeList> result = getMNUsageBySnapshot(llmqParameters.getType(), quorumBaseBlock, snapshot);
-        SimplifiedMasternodeList mnsUsedAtH = result.getFirst();
-        SimplifiedMasternodeList mnsNotUsedAtH = result.getSecond();
-
-        ArrayList<Masternode> sortedMnsUsedAtH = mnsUsedAtH.calculateQuorum(mnsUsedAtH.getAllMNsCount(), modifier);
-
-        ArrayList<Masternode> sortedMnsNotUsedAtH = mnsNotUsedAtH.calculateQuorum(mnsNotUsedAtH.getAllMNsCount(), modifier);
-        ArrayList<Masternode> sortedCombinedMnsList = new ArrayList<>(sortedMnsNotUsedAtH);
-        for (Masternode m1 : sortedMnsUsedAtH) {
-            for (Masternode m2: sortedMnsNotUsedAtH) {
-                if (m1.equals(m2)) {
-                    log.info("{} is in both lists", m1);
-                }
-            }
-        }
-        sortedCombinedMnsList.addAll(sortedMnsUsedAtH);
-
-        //Mode 0: No skipping
-        if (snapshot.getSkipListMode() == SnapshotSkipMode.MODE_NO_SKIPPING.getValue()){
-            Iterator<Masternode> itm = sortedCombinedMnsList.iterator();
-            for (int i = 0; i < llmqParameters.getSigningActiveQuorumCount(); ++i) {
-                //Iterate over the first quarterSize elements
-                while (quarterQuorumMembers.get(i).size() < quarterSize) {
-                    Masternode m = itm.next();
-                    quarterQuorumMembers.get(i).add((SimplifiedMasternodeListEntry) m);
-                    if (!itm.hasNext()) {
-                        itm = sortedCombinedMnsList.iterator();
-                    }
-                }
-            }
-        }
-        //Mode 1: List holds entries to be skipped
-        else if (snapshot.getSkipListMode() == SnapshotSkipMode.MODE_SKIPPING_ENTRIES.getValue()){
-            HashSet<Sha256Hash> mnProTxHashToRemove = new HashSet<>();
-            int first_entry_index = 0;
-            ArrayList<Integer> processedSkipList = Lists.newArrayList();
-            for (int s : snapshot.getSkipList()) {
-                if (first_entry_index == 0) {
-                    first_entry_index = s;
-                    processedSkipList.add(s);
-                }
-                else {
-                    processedSkipList.add(first_entry_index + s);
-                }
-            }
-
-            int idx = 0;
-            Iterator<Integer> itsk = processedSkipList.iterator();
-            for (int i = 0; i < llmqParameters.getSigningActiveQuorumCount(); ++i) {
-                //Iterate over the first quarterSize elements
-                while (quarterQuorumMembers.get(i).size() < quarterSize) {
-                    if (itsk.hasNext() && idx == itsk.next())
-                        ;
-                    else
-                        quarterQuorumMembers.get(i).add((SimplifiedMasternodeListEntry) sortedCombinedMnsList.get(idx));
-                    idx++;
-                    if (idx == sortedCombinedMnsList.size())
-                        idx = 0;
-                }
-            }
-        }
-        //Mode 2: List holds entries to be kept
-        else if (snapshot.getSkipListMode() == SnapshotSkipMode.MODE_NO_SKIPPING_ENTRIES.getValue()) {
-            // Mode 2 will be written later
-        }
-        //Mode 3: Every node was skipped. Returning empty quarterQuorumMembers
-
-        return quarterQuorumMembers;
-    }
-
-    Pair<SimplifiedMasternodeList, SimplifiedMasternodeList> getMNUsageBySnapshot(LLMQParameters.LLMQType llmqType, StoredBlock quorumBaseBlock, QuorumSnapshot snapshot)
-    {
-        SimplifiedMasternodeList usedMNs = new SimplifiedMasternodeList(params);
-        SimplifiedMasternodeList nonUsedMNs = new SimplifiedMasternodeList(params);
-
-        SimplifiedMasternodeList mns = getListForBlock(quorumBaseBlock.getHeader().getHash());
-        ArrayList<Masternode> list = mns.calculateQuorum(mns.getValidMNsCount(), quorumBaseBlock.getHeader().getHash());
-
-        AtomicInteger i = new AtomicInteger();
-
-        for (Masternode mn: list) {
-            if (snapshot.getActiveQuorumMembers().get(i.get())) {
-                usedMNs.addMN((SimplifiedMasternodeListEntry) mn);
-            } else {
-                nonUsedMNs.addMN((SimplifiedMasternodeListEntry) mn);
-            }
-            i.getAndIncrement();
-        }
-
-        return new Pair<>(usedMNs, nonUsedMNs);
     }
 
     public boolean isSynced() {
